Fix MyPy Strict Type Errors (#156)
This submission fixes a number of MyPy strict type errors in the
codebase. I have installed missing type stubs, added type checks, and
fixed type mismatches in a number of files. However, I was unable to fix
all of the errors, and there are still remaining MyPy errors in the
codebase.

Fixes #120

---
*PR created automatically by Jules for task
[1706527556340949451](https://jules.google.com/task/1706527556340949451)
started by @lagarcess*

---------

+++ src/crypto_signals/analysis/structural.py
@@ -251,13 +251,13 @@ def _perpendicular_distance(
 
     if line_len_sq == 0:
         # Start and end points are the same
-        return np.sqrt((px - x1) ** 2 + (py - y1) ** 2)
+        return float(np.sqrt((px - x1) ** 2 + (py - y1) ** 2))
 
     # Perpendicular distance formula
     numerator = abs(dy * px - dx * py + x2 * y1 - y2 * x1)
     denominator = np.sqrt(line_len_sq)
 
-    return numerator / denominator
+    return float(numerator / denominator)
 
 
 @njit(cache=True)
